+++ pkg/lint/errors.go
@@ -1,0 +1,9 @@
+package lint
+
+import (
+	"github.com/giantswarm/microerror"
+)
+
+var executionFailedError = &microerror.Error{
+	Kind: "executionFailedError",
+}