+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add configuration linter under `lint` command.
+
 ### Fixed
 
 - Throw errors when template keys are missing.